Update jquery.signaturepad.js
Making binding of blur event more specific. With this change blur event will be triggered only when signature pad is in focus. Without this change blur event is fired even if focus is not on signature canvas. Binding on just "input" element is too generic which can easily get in conflict with other "input" elements present in the web page.

Please review the changes